Update insert_glitch.py
Finished. Can now plot efficiencies of an entire TOD. I can't figure out how to get 'ld' to compute the way we've been doing it without making the cuts before we load the TOD (impossible). For now this works. Just keep it as ids[2000] or else the detector list doesn't work.